[wip] Docs, models, and more for Reports API.
