+++ docs/requirements.txt
@@ -1,3 +1,3 @@
-sphinx>=3.2.0,<3.3
+sphinx>=3.4.3,<3.5
 sphinx-autobuild>=0.7.1
 sphinx-rtd-theme
