+++ docs/source/conf.py
@@ -42,6 +42,7 @@ extensions = [
 
 # Default "alphabetical", change to source ordering:
 autodoc_member_order = "bysource"
+autodoc_typehints = "description"
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ["_templates"]
